Fixed warnings.

+++ trunk/gui.c
@@ -1219,7 +1219,7 @@ void gui_render_perf( struct ahx_tune *at, struct ahx_instrument *ins, BOOL forc
     if( n > 60 ) n = 0;
     
     sprintf( pbuf, "%03d %s%c %01d %01X%02X %01X%02X ",
-      j,
+      (int)j,
       notenames[n],
       ins->ins_PList.pls_Entries[j].ple_Fixed ? '*' : ' ',
       ins->ins_PList.pls_Entries[j].ple_Waveform,
@@ -1308,7 +1308,7 @@ void gui_render_inslistb( BOOL force )
     set_pens(&bitmaps[BM_INSLISTB], PAL_TEXT, PAL_BACK);
     for( i=0, j=at->at_topinsb; i<(INSLSTB_H>>4); i++, j++ )
     {
-      sprintf( tmp, "%02d                   ", j );
+      sprintf( tmp, "%02d                   ", (int)j );
       in = &at->at_Instruments[j];
 
       for( k=0; k<18; k++ )
@@ -2091,7 +2091,7 @@ void gui_render_tracked( BOOL force )
      
       if( ( i >= 0 ) && ( i < trklen ) )
       {
-        sprintf( pbuf, "%02d ", i );
+        sprintf( pbuf, "%02d ", (int)i );
         for( k=lch, l=0; k<ech; k++, l++ )
         {
           stp = &at->at_Tracks[at->at_Positions[posnr].pos_Track[k]&0xff][i];
@@ -2157,7 +2157,7 @@ void gui_render_tracked( BOOL force )
         { 
           if( k < ech )
           {
-            sprintf( pbuf, "Track %d", k+1 );
+            sprintf( pbuf, "Track %d", (int)k+1 );
             set_fpen(&bitmaps[BM_TRACKBAR], PAL_BTNSHADOW);
             printstr(&bitmaps[BM_TRACKBAR], pbuf, j+1, 2);
             set_fpen(&bitmaps[BM_TRACKBAR], PAL_BTNTEXT);
@@ -2222,7 +2222,7 @@ void gui_render_tracked( BOOL force )
 
         if( ( i >= 0 ) && ( i < trklen ) )
         {
-          sprintf( pbuf, "%02d ", i );
+          sprintf( pbuf, "%02d ", (int)i );
           for( k=lch, l=0; k<ech; k++, l++ )
           {
             stp = &at->at_Tracks[at->at_Positions[posnr].pos_Track[k]&0xff][i];
@@ -2427,7 +2427,7 @@ void gui_render_inslist( BOOL force )
     set_pens(&bitmaps[BM_INSLIST], PAL_TEXT, PAL_BACK);
     for( i=0, j=at->at_topins; i<9; i++, j++ )
     {
-      sprintf( tmp, "%02d                      ", j );
+      sprintf( tmp, "%02d                      ", (int)j );
       in = &at->at_Instruments[j];
 
       for( k=0; k<21; k++ )
@@ -2540,7 +2540,7 @@ void gui_render_posed( BOOL force )
      
       if( ( i >= 0 ) && ( i < 1000 ) )
       {
-        sprintf( pbuf, "%03d ", i );
+        sprintf( pbuf, "%03d ", (int)i );
         for( k=lch, l=0; k<ech; k++, l++ )
           sprintf( &pbuf[l*7+4], "%03d %02X ",
             at->at_Positions[i].pos_Track[k]&0xff,
@@ -2626,7 +2626,7 @@ void gui_render_posed( BOOL force )
     set_pens(&bitmaps[BM_POSED], PAL_POSEDCHIND, PAL_BARMID);
     for( k=lch, l=0; k<ech; k++, l++ )
     {
-      sprintf( pbuf, "%d ", k+1 );
+      sprintf( pbuf, "%d ", (int)k+1 );
       printstr(&bitmaps[BM_POSED], pbuf, l*49+25, 0);
     }      
     set_bpen(&bitmaps[BM_POSED], PAL_BACK);
